+++ packages/backend/.env.example
@@ -2,3 +2,4 @@ TRELLO_KEY=
 TRELLO_TOKEN=
 
 TRELLO_BOARD_ID=
+TRELLO_UPDATE_CRON=0 0 */1 * * *
