+++ packages/backend/package.json
@@ -13,10 +13,12 @@
 	},
 	"dependencies": {
 		"axios": "^0.19.2",
-		"dotenv": "^8.2.0"
+		"dotenv": "^8.2.0",
+		"node-cron": "^2.0.3"
 	},
 	"devDependencies": {
 		"@types/node": "^12.12.31",
+		"@types/node-cron": "^2.0.3",
 		"ts-node-dev": "^1.0.0-pre.44"
 	}
 }
